+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-**UNRELEASED**
+**2.0.1**
 
 - Fixed broken ``MultiListener.extra()`` (PR by daa)
 - Fixed ``TLSStream`` returning an empty bytes object instead of raising ``EndOfStream`` when